requirements.txt
flask
flask_sqlalchemy
flask_migrate
flask_jwt_extended
psycopg2-binary
marshmallow
python-dotenv
flask-swagger-ui
flask_smorest
flask-cors